refactor: rename port types module
