+++ hippoium/core/cer/runtime.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 from collections import deque
 from statistics import fmean
-from hippoium.ports.types import TrimPolicy
+from hippoium.ports.port_types import TrimPolicy
 from hippoium.ports.constants import MAX_TOKENS_PROMPT
 from hippoium.ports.protocols import TokenMeter
 from hippoium.core.utils.token_counter import count_tokens
