+++ hippoium/core/negative/auto_labeler.py
@@ -4,7 +4,7 @@ Auto-label incoming/outgoing messages as OK/WARN/ERR.
 from __future__ import annotations
 import re
 from hippoium.ports.schemas import Message
-from hippoium.ports.types import MsgLabel
+from hippoium.ports.port_types import MsgLabel
 
 
 _ERR_PATTERNS = [
